Feat: build and style teh order summary section
