+++ src/App.tsx
@@ -6,13 +6,11 @@ import {
 } from 'react-router-dom';
 
 import Checkout from './pages/Checkout';
-import Product from './pages/Product';
 
 const router = createBrowserRouter(
   createRoutesFromElements(
     <>
-      <Route path="/" element={<Product />} />
-      <Route path="/checkout" element={<Checkout />} />
+      <Route path="/" element={<Checkout />} />
     </>
   )
 );
