+++ src/components/NavBar/NavBar.scss
@@ -19,13 +19,16 @@ header {
     > ul {
       @include display(row, unset, center);
 
-      gap: 1rem;
+      gap: 1.3rem;
 
       li {
         @include font-stack(0.9rem, 500, 24px);
 
         a {
+          @include display(row, unset, center);
+
           color: #fff;
+          gap: 5px;
 
           &:hover {
             border-bottom: solid 2px $acc-color-1;
