+++ src/components/Product/ProductItem.tsx
@@ -3,11 +3,12 @@ const Product = ({ product, addProductToCart, removeFromCart }: any) => {
     <div className="product-card">
       <h2>{product.name}</h2>
       <p>${product.price}</p>
+      <p>Available Quantity: <span>{product.quantity}</span></p>
       <p>{product.description}</p>
       <div>
         <button onClick={addProductToCart}>Add</button>
         <p>
-          Qty: <span>{product.quantity}</span>
+          <span>{product.quantity}</span>
         </p>
         <button onClick={removeFromCart}>Remove</button>
       </div>
