+++ src/styles/_variables.scss
@@ -1,13 +1,14 @@
 
 @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500;600;700;800;900&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Raleway:wght@100;200;300;400;500;600;700;800;900&display=swap');
 
-$pri-color-1: #2B7CCF;
+$pri-color-1: #B9EDDD;
 $pri-color-2: #435faa;
 $pri-color-3: #F2F5FB;
 $pri-color-4: #ebf1fc;
 $pri-color-5: #d2def6;
 
-$acc-color-1: #6fed6c;
+$acc-color-1: #577D86;
 $acc-color-2: #3ebe98;
 
 $tet-color-3: #F2B925;
@@ -16,10 +17,11 @@ $tet-color-4: #F26F25;
 $neu-color-1: #ecf9fb;
 $neu-color-2: #B8BFD1;
 
-$dark: rgb(52, 51, 51);
+$dark: #44536e;
 
 $shadow-1: rgba(0, 0, 0, 0.25) 0 25px 50px -12px;
 $shadow-2: rgba(50, 50, 93, 0.25) 0 6px 12px -2px, rgba(0, 0, 0, 0.3) 0 3px 7px -3px;
 $shadow-3: rgba(17, 12, 46, 0.15) 0 48px 100px 0;
 
-$font-1: 'Poppins', sans-serif;
+$font-1: 'Raleway', sans-serif;
+$font-2: 'Poppins', sans-serif;
